# api.yaml
openapi: 3.0.0
info:
  title: API de Gestión de Clientes y Productos
  description: API para realizar operaciones CRUD en clientes y productos.
  version: 1.0.0
servers:
  - url: http://localhost:3000/
paths:
  /clientes:
    get:
      summary: Listar Clientes
      description: Obtiene una lista de todos los clientes.
      responses:
        '200':
          description: Lista de clientes exitosamente recuperada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
    post:
      summary: Crear Cliente
      description: Crea un nuevo cliente.
      requestBody:
        description: Datos del cliente a crear.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseCliente'
      responses:
        '201':
          description: Cliente creado exitosamente.
  /clientes/{nro_cliente}:
    parameters:
      - name: nro_cliente
        in: path
        required: true
        description: Número de cliente.
        schema:
          type: integer
    get:
      summary: Obtener Cliente por ID
      description: Obtiene un cliente por su número de cliente.
      responses:
        '200':
          description: Cliente recuperado exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado.
    put:
      summary: Actualizar Cliente por ID
      description: Actualiza un cliente existente.
      requestBody:
        description: Datos actualizados del cliente.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cliente'
      responses:
        '200':
          description: Cliente actualizado exitosamente.
        '404':
          description: Cliente no encontrado.
    delete:
      summary: Eliminar Cliente por ID
      description: Elimina un cliente por su número de cliente.
      responses:
        '204':
          description: Cliente eliminado exitosamente.
        '404':
          description: Cliente no encontrado.
...
components:
  schemas:
    BaseCliente:
      type: object
      properties:
        nombre:
          type: string
          example: Alan
        apellido:
          type: string
          example: Martinez
        direccion:
          type: string
          example: Calle 123412
        activo:
          type: integer
          example: 12
        telefonos:
          type: array
          items:
            type: object
            properties:
              tipo:
                type: string
                enum: [F, M]
                example: 'F'
              nro_telefono:
                type: integer
                example: 124141411
              codigo_area:
                type: integer
                example: 135
    Cliente:
      allOf:
        - $ref: '#/components/schemas/BaseCliente'
        - type: object
          properties:
            nro_cliente:
              type: integer
...